Refactor single profile router
This was doing a weird thing where sometimes a request to '/' was wanting a list, and sometimes it wanted a single profile.

Fix that so a request to '/' on the profle router is always a list, and create a separate router for a single profile that can be mounted in the contexts where it is needed.
