+++ lib/routers/establishment/index.js
@@ -59,10 +59,6 @@ router.get('/:establishment', (req, res, next) => {
 router.use('/:establishment/roles', require('./roles'));
 router.use('/:establishment/place(s)?', require('./places'));
 
-router.use('/:establishment/profile(s)?/:profileId', (req, res, next) => {
-  req.profileId = req.params.profileId;
-  next();
-});
 router.use('/:establishment/profile(s)?', require('../profile'));
 
 router.use('/:establishment/project(s)?', require('./projects'));
