+++ lib/routers/profile/index.js
@@ -1,74 +1,5 @@
-const { omit } = require('lodash');
 const { Router } = require('express');
-const isUUID = require('uuid-validate');
-const { permissions } = require('../../middleware');
-const { validateSchema } = require('../../middleware');
 const { NotFoundError } = require('../../errors');
-
-const update = () => (req, res, next) => {
-  const params = {
-    model: 'profile',
-    id: req.profileId,
-    data: req.body.data || req.body,
-    meta: req.body.meta
-  };
-
-  req.workflow.update(params)
-    .then(response => {
-      res.response = response;
-      next();
-    })
-    .catch(next);
-};
-
-const validate = () => {
-  return (req, res, next) => {
-    const ignoredFields = ['comments'];
-    return validateSchema(req.models.Profile, {
-      ...(req.user.profile || {}),
-      ...omit(req.body, ignoredFields)
-    })(req, res, next);
-  };
-};
-
-const getSingleProfile = req => {
-  if (!isUUID(req.profileId)) {
-    throw new NotFoundError();
-  }
-  const { Profile } = req.models;
-  const profile = Profile.scopeSingle({
-    id: req.profileId,
-    userId: req.user.profile.id,
-    establishmentId: (req.establishment && req.establishment.id) || undefined
-  });
-
-  // own profile
-  if (req.profileId === req.user.profile.id) {
-    return profile.get();
-  }
-
-  return Promise.resolve()
-    .then(() => req.user.can('profile.read.all', req.params))
-    .then(allowed => {
-      if (allowed) {
-        return profile.get();
-      }
-      return Promise.resolve()
-        .then(() => req.user.can('profile.read.basic', req.params))
-        .then(allowed => {
-          if (allowed) {
-            return profile.getNamed();
-          }
-          throw new NotFoundError();
-        });
-    })
-    .then(profile => {
-      if (!profile) {
-        throw new NotFoundError();
-      }
-      return profile;
-    });
-};
 
 const getAllProfiles = req => {
   const { Profile } = req.models;
@@ -103,16 +34,7 @@ const getAllProfiles = req => {
 
 const router = Router({ mergeParams: true });
 
-router.use((req, res, next) => {
-  if (req.profileId) {
-    return Promise.resolve()
-      .then(() => getSingleProfile(req))
-      .then(profile => {
-        res.response = profile;
-        next();
-      })
-      .catch(next);
-  }
+router.get('/', (req, res, next) => {
   Promise.resolve()
     .then(() => getAllProfiles(req))
     .then(({ filters, total, profiles }) => {
@@ -125,17 +47,16 @@ router.use((req, res, next) => {
     .catch(next);
 });
 
-router.put('/', validate(), update());
+router.param('profileId', (req, res, next, profileId) => {
+  req.profileId = profileId;
+  next();
+});
 
-router.put('/:id',
-  permissions('profile.update'),
-  validate(),
-  update()
-);
+router.use('/:profileId', require('./person'));
 
-router.use('/:id/certificate', require('./certificates'));
-router.use('/:id/exemption', require('./exemptions'));
-router.use('/:id/pil', require('./pil'));
-router.use('/:id/permission', require('./permission'));
+router.use('/:profileId/certificate', require('./certificates'));
+router.use('/:profileId/exemption', require('./exemptions'));
+router.use('/:profileId/pil', require('./pil'));
+router.use('/:profileId/permission', require('./permission'));
 
 module.exports = router;
