+++ lib/routers/user.js
@@ -3,11 +3,6 @@ const { Router } = require('express');
 const router = Router();
 
 router.use((req, res, next) => {
-  req.profileId = req.user.profile.id;
-  next();
-});
-
-router.use((req, res, next) => {
   Promise.resolve()
     .then(() => req.user.allowedActions())
     .then(allowedActions => {
@@ -17,7 +12,12 @@ router.use((req, res, next) => {
     .catch(next);
 });
 
-router.use(require('./profile'));
+router.use((req, res, next) => {
+  req.profileId = req.user.profile.id;
+  next();
+});
+
+router.use(require('./profile/person'));
 
 router.use((req, res, next) => {
   const { Invitation } = req.models;
